+++ examples/usbpower/test.py
@@ -1,0 +1,7 @@
+from labgrid import Environment
+from labgrid.protocol import ConsoleProtocol
+
+env = Environment('example-env.yaml')
+cp = t.get_driver(ConsoleProtocol)
+
+